+++ library/src/main/java/io/sfrei/tracksearch/clients/youtube/YouTubeClient.java
@@ -1,7 +1,9 @@
 package io.sfrei.tracksearch.clients.youtube;
 
+import io.sfrei.tracksearch.clients.helper.ClientHelper;
 import io.sfrei.tracksearch.clients.setup.*;
 import io.sfrei.tracksearch.config.TrackSearchConfig;
+import io.sfrei.tracksearch.config.TrackSearchConstants;
 import io.sfrei.tracksearch.exceptions.TrackSearchException;
 import io.sfrei.tracksearch.exceptions.YouTubeException;
 import io.sfrei.tracksearch.tracks.BaseTrackList;
@@ -10,7 +12,7 @@ import io.sfrei.tracksearch.tracks.TrackList;
 import io.sfrei.tracksearch.tracks.YouTubeTrack;
 import io.sfrei.tracksearch.tracks.metadata.YouTubeTrackFormat;
 import io.sfrei.tracksearch.tracks.metadata.YouTubeTrackInfo;
-import io.sfrei.tracksearch.utils.ScriptCache;
+import io.sfrei.tracksearch.utils.CacheMap;
 import io.sfrei.tracksearch.utils.TrackFormatUtility;
 import io.sfrei.tracksearch.utils.TrackListHelper;
 import io.sfrei.tracksearch.utils.URLUtility;
@@ -25,12 +27,12 @@ import java.util.Map;
 @Slf4j
 public class YouTubeClient extends SingleSearchClient<YouTubeTrack> {
 
-    public static final String HOSTNAME = "https://youtube.com";
+    public static final String HOSTNAME = "https://www.youtube.com";
+    public static final String PAGING_KEY = "ctoken";
     private static final String INFORMATION_PREFIX = "yt";
     public static final String POSITION_KEY = INFORMATION_PREFIX + TrackSearchConfig.POSITION_KEY_SUFFIX;
     public static final String OFFSET_KEY = INFORMATION_PREFIX + TrackSearchConfig.OFFSET_KEY_SUFFIX;
     private static final String PAGING_INFORMATION = INFORMATION_PREFIX + "PagingToken";
-    public static final String PAGING_KEY = "ctoken";
     private static final String ADDITIONAL_PAGING_KEY = "continuation";
 
     private static final Map<String, String> VIDEO_SEARCH_PARAMS = Map.of("sp", "EgIQAQ%3D%3D");
@@ -47,9 +49,15 @@ public class YouTubeClient extends SingleSearchClient<YouTubeTrack> {
     private final YouTubeService requestService;
     private final YouTubeUtility youTubeUtility;
 
-    private final ScriptCache<String, String> scriptCache;
+    private final CacheMap<String, String> scriptCache;
 
     public YouTubeClient() {
+
+        super(
+                (uri, cookie) -> cookie.getName().equals("YSC") || cookie.getName().equals("VISITOR_INFO1_LIVE") || cookie.getName().equals("GPS"),
+                Map.of("Cookie", "CONSENT=YES+cb.20210328-17-p0.en+FX+478")
+        );
+
         final Retrofit base = new Retrofit.Builder()
                 .baseUrl(HOSTNAME)
                 .client(okHttpClient)
@@ -58,7 +66,7 @@ public class YouTubeClient extends SingleSearchClient<YouTubeTrack> {
 
         requestService = base.create(YouTubeService.class);
         youTubeUtility = new YouTubeUtility();
-        scriptCache = new ScriptCache<>();
+        scriptCache = new CacheMap<>();
     }
 
     public static Map<String, String> makeQueryInformation(final String query, final String pagingToken) {
@@ -83,9 +91,9 @@ public class YouTubeClient extends SingleSearchClient<YouTubeTrack> {
 
     private String provideStreamUrl(final YouTubeTrack track) {
         try {
-            return getStreamUrl(track);
+            return getStreamUrl(track, TrackSearchConstants.RETRY_RESOLVING_ONCE);
         } catch (TrackSearchException e) {
-            log.error(e.getMessage());
+            log.error("Error occurred acquiring stream URL", e);
         }
         return null;
     }
@@ -105,41 +113,49 @@ public class YouTubeClient extends SingleSearchClient<YouTubeTrack> {
         throw new YouTubeException("Query type not supported");
     }
 
-    private YouTubeTrackInfo loadTrackInfo(final YouTubeTrack youtubeTrack) throws TrackSearchException {
+    public YouTubeTrackInfo loadTrackInfo(final YouTubeTrack youtubeTrack) throws TrackSearchException {
         final String trackUrl = youtubeTrack.getUrl();
         final Call<ResponseWrapper> trackRequest = requestService.getForUrlWithParams(trackUrl, TRACK_PARAMS);
         final ResponseWrapper trackResponse = Client.request(trackRequest);
 
         final String trackContent = trackResponse.getContentOrThrow();
-        final YouTubeTrackInfo trackInfo = youTubeUtility.getTrackInfo(trackContent, trackUrl);
+        final YouTubeTrackInfo trackInfo = youTubeUtility.getTrackInfo(trackContent, trackUrl, this::requestURL);
         return youtubeTrack.setAndGetTrackInfo(trackInfo);
     }
 
     @Override
     public String getStreamUrl(@NonNull final YouTubeTrack youtubeTrack) throws TrackSearchException {
+
         final YouTubeTrackInfo trackInfo = loadTrackInfo(youtubeTrack);
 
         final YouTubeTrackFormat youtubeTrackFormat = TrackFormatUtility.getBestTrackFormat(youtubeTrack, false);
 
         if (youtubeTrackFormat.isStreamReady())
-            return youtubeTrackFormat.getUrl();
+            return URLUtility.decode(youtubeTrackFormat.getUrl());
 
         final String scriptUrl = trackInfo.getScriptUrl();
         if (scriptUrl == null)
             throw new TrackSearchException("ScriptURL could not be resolved");
 
         final String scriptContent;
-        if (scriptCache.containsKey(scriptUrl))
+        if (scriptCache.containsKey(scriptUrl)) {
+            log.trace("Use cached script for: {}", scriptUrl);
             scriptContent = scriptCache.get(scriptUrl);
-        else {
-            scriptContent = Client.requestURL(HOSTNAME + scriptUrl).getContentOrThrow();
+        } else {
+            scriptContent = requestURL(HOSTNAME + scriptUrl).getContentOrThrow();
             scriptCache.put(scriptUrl, scriptContent);
         }
 
-        final String signatureKey = youTubeUtility.getSignature(youtubeTrackFormat, scriptContent);
+        final String signature = youTubeUtility.getSignature(youtubeTrackFormat, scriptUrl, scriptContent);
+        final String unauthorizedStreamUrl = youtubeTrackFormat.getUrl();
 
-        final String streamUrl = youtubeTrackFormat.getUrl();
-        return URLUtility.appendParam(streamUrl, youtubeTrackFormat.getSigParam(), signatureKey);
+        return URLUtility.addRequestParam(unauthorizedStreamUrl, youtubeTrackFormat.getSigParam(), signature);
+    }
+
+    @Override
+    public String getStreamUrl(@NonNull final YouTubeTrack youtubeTrack, final int retries) throws TrackSearchException {
+        return ClientHelper.getStreamUrl(this, youtubeTrack, this::requestAndGetCode, retries)
+                .orElseThrow(() -> new YouTubeException(String.format("Not able to get stream URL after %s retries", retries)));
     }
 
     private Map<String, String> getPagingParams(final Map<String, String> queryInformation) {