+++ library/src/main/java/io/sfrei/tracksearch/utils/TimeUtility.java
@@ -1,16 +1,19 @@
 package io.sfrei.tracksearch.utils;
 
+import lombok.experimental.UtilityClass;
+
 import java.text.SimpleDateFormat;
 import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
 
+@UtilityClass
 public class TimeUtility {
 
-    private static final Date DATE_REFERENCE = new Date(0);
-    private static final List<String> TIME_FORMATS = Arrays.asList("HH:mm:ss", "mm:ss", "ss");
+    private final Date DATE_REFERENCE = new Date(0);
+    private final List<String> TIME_FORMATS = Arrays.asList("HH:mm:ss", "mm:ss", "ss");
 
-    public static Long getSecondsForTimeString(final String time) {
+    public Long getSecondsForTimeString(final String time) {
         for (final String format : TIME_FORMATS) {
             try {
                 return (new SimpleDateFormat(format).parse(time).getTime() - DATE_REFERENCE.getTime()) / 1000L;
@@ -20,7 +23,7 @@ public class TimeUtility {
         return null;
     }
 
-    public static Long getSecondsForMilliseconds(final Long milliseconds) {
+    public Long getSecondsForMilliseconds(final Long milliseconds) {
         if (milliseconds == null)
             return null;